+++ CHANGES.md
@@ -26,6 +26,9 @@
    * Changed the String/List versions of `.++/.!!` to directly use the names `++/!!`. Since these modules
      are intended to be used qualified only, there's no reason to add the dots.
 
+   * Added `Documentation.SBV.Examples.Lists.CountOutAndTransfer` example, proving that the so-called
+     coating card trick works correctly.
+
 ### Version 8.17, 2021-10-25
 
   * SBV now supports cvc5; the latest incarnation of CVC. See https://github.com/cvc5/cvc5